ncd_app/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.8
requests==2.32.3
django-cors-headers==4.4.0
